requirements.txt
langchain>=0.1.0
langgraph>=0.0.15
gradio>=4.19.2
fastapi>=0.109.0
uvicorn>=0.27.0
python-dotenv>=1.0.0
pydantic>=2.6.1
openai>=1.12.0
typing-extensions>=4.9.0
python-multipart>=0.0.6
pytest>=8.0.0
pytest-asyncio>=0.23.5
httpx>=0.26.0
pytest-cov>=4.1.0 